src/agents/base/pyproject.toml
[project]
name = "base-agent"
version = "1.0.0"
description = "General purpose AI assistant agent"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.0",
    "httpx>=0.28.1",
    "uvicorn>=0.34.0",
    "starlette>=0.46.1",
    "pydantic>=2.11.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "openai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
